[FEAT] Split out marketplace components
