+++ src/features/marketplace/components/Tradeable.tsx
@@ -1,27 +1,15 @@
-import { Loading } from "features/auth/components";
 import {
   CollectionName,
   TradeableDetails,
 } from "features/game/types/marketplace";
 import React, { useContext, useEffect, useState } from "react";
 import * as Auth from "features/auth/lib/Provider";
-import { useActor, useSelector } from "@xstate/react";
+import { useActor } from "@xstate/react";
 import { useNavigate, useParams } from "react-router-dom";
 import { loadTradeable } from "../actions/loadTradeable";
-import { InnerPanel, Panel } from "components/ui/Panel";
-import { getTradeableDisplay, TradeableDisplay } from "../lib/tradeables";
+import { getTradeableDisplay } from "../lib/tradeables";
 
-import bg from "assets/ui/3x3_bg.png";
-import walletIcon from "assets/icons/wallet.png";
-import sflIcon from "assets/icons/sfl.webp";
-import tradeIcon from "assets/icons/trade.png";
-
-import { Label } from "components/ui/Label";
-import { SUNNYSIDE } from "assets/sunnyside";
-import { Button } from "components/ui/Button";
 import { PriceHistory } from "./PriceHistory";
-import { TradeTable } from "./TradeTable";
-import { useAppTranslation } from "lib/i18n/useAppTranslations";
 import { TradeableOffers, YourOffer } from "./TradeableOffers";
 import { Context } from "features/game/GameProvider";
 import { KNOWN_ITEMS } from "features/game/types";
@@ -31,9 +19,9 @@ import {
 } from "features/island/hud/components/inventory/utils/inventory";
 import { ITEM_NAMES } from "features/game/types/bumpkin";
 import { availableWardrobe } from "features/game/events/landExpansion/equip";
-import { TradeableListItem } from "./TradeableList";
-import { Modal } from "components/ui/Modal";
-import { MachineState } from "features/game/lib/gameMachine";
+import { TradeableHeader } from "./TradeableHeader";
+import { TradeableInfo } from "./TradeableInfo";
+import { TradeableListings } from "./TradeableListings";
 
 export const Tradeable: React.FC = () => {
   const { authService } = useContext(Auth.Context);
@@ -165,225 +153,5 @@ export const Tradeable: React.FC = () => {
         />
       </div>
     </div>
-  );
-};
-
-const TradeableHeader: React.FC<{
-  collection: CollectionName;
-  display: TradeableDisplay;
-  tradeable?: TradeableDetails;
-  count: number;
-  onBack: () => void;
-  onListClick: () => void;
-}> = ({ collection, onBack, display, count, tradeable, onListClick }) => {
-  const { t } = useAppTranslation();
-
-  const cheapestListing = tradeable?.listings.reduce((cheapest, listing) => {
-    return listing.sfl < cheapest.sfl ? listing : cheapest;
-  }, tradeable?.listings?.[0]);
-
-  return (
-    <InnerPanel className="w-full mb-1">
-      <div className="p-2">
-        <div className="flex flex-wrap items-center justify-between">
-          <div
-            className="flex cursor-pointer items-center w-fit mb-2"
-            onClick={onBack}
-          >
-            <img src={SUNNYSIDE.icons.arrow_left} className="h-6 mr-2" />
-            <p className="capitalize underline">{collection}</p>
-          </div>
-          <div className="flex items-center">
-            {tradeable?.type === "onchain" && (
-              <Label type="formula" className="mr-2" icon={walletIcon}>
-                {t("marketplace.walletRequired")}
-              </Label>
-            )}
-
-            {tradeable !== undefined && (
-              <Label type="default">{t("marketplace.youOwn", { count })}</Label>
-            )}
-          </div>
-        </div>
-        <div className="flex">
-          <p className="text-lg mr-0.5 mb-1">{display.name}</p>
-        </div>
-        <div className="flex items-center justify-between flex-wrap">
-          {cheapestListing ? (
-            <div className="flex items-center mr-2 sm:mb-0.5 -ml-1">
-              <>
-                <img src={sflIcon} className="h-8 mr-2" />
-                <p className="text-base">{`${cheapestListing.sfl} SFL`}</p>
-              </>
-            </div>
-          ) : (
-            // Dummy div to keep the layout consistent
-            <div className="flex items-center mr-2 sm:mb-0.5 -ml-1"></div>
-          )}
-          <div className="items-center justify-between hidden sm:flex sm:visible w-full sm:w-auto">
-            {cheapestListing && (
-              <Button className="mr-1 w-full sm:w-auto">
-                {t("marketplace.buyNow")}
-              </Button>
-            )}
-            <Button
-              onClick={onListClick}
-              disabled={!count}
-              className="w-full sm:w-auto"
-            >
-              {t("marketplace.listForSale")}
-            </Button>
-          </div>
-        </div>
-      </div>
-      <div className="flex items-center justify-between sm:hidden w-full sm:w-auto">
-        {cheapestListing && (
-          <Button className="mr-1 w-full sm:w-auto">
-            {t("marketplace.buyNow")}
-          </Button>
-        )}
-        <Button
-          onClick={onListClick}
-          disabled={!count}
-          className="w-full sm:w-auto"
-        >
-          {t("marketplace.listForSale")}
-        </Button>
-      </div>
-    </InnerPanel>
-  );
-};
-
-const TradeableInfo: React.FC<{
-  display: TradeableDisplay;
-  tradeable?: TradeableDetails;
-}> = ({ display, tradeable }) => {
-  const { t } = useAppTranslation();
-
-  return (
-    <>
-      <InnerPanel className="w-full flex relative mb-1">
-        {tradeable && (
-          <Label
-            className="absolute top-2 right-2"
-            type="default"
-          >{`Supply: ${tradeable.supply}`}</Label>
-        )}
-
-        <img src={bg} className="w-full rounded-sm" />
-        <img
-          src={display.image}
-          className="w-1/3 absolute"
-          style={{
-            left: "50%",
-            transform: "translate(-50%, 50%)",
-            bottom: "50%",
-          }}
-        />
-      </InnerPanel>
-      <InnerPanel>
-        <div className="p-2">
-          <Label type="default" className="mb-1" icon={SUNNYSIDE.icons.search}>
-            {t("marketplace.description")}
-          </Label>
-          <p className="text-sm mb-2">{display.description}</p>
-          {display.buff && (
-            <Label
-              icon={display.buff.boostTypeIcon}
-              type={display.buff.labelType}
-              className="mb-2"
-            >
-              {display.buff.shortDescription}
-            </Label>
-          )}
-        </div>
-      </InnerPanel>
-    </>
-  );
-};
-
-const _isListing = (state: MachineState) => state.matches("effect");
-
-const TradeableListings: React.FC<{
-  tradeable?: TradeableDetails;
-  display: TradeableDisplay;
-  farmId: number;
-  collection: CollectionName;
-  id: number;
-  showListItem: boolean;
-  count: number;
-  onListClick: () => void;
-  onListClose: () => void;
-  onListingMade: () => void;
-}> = ({
-  tradeable,
-  farmId,
-  display,
-  id,
-  count,
-  showListItem,
-  onListingMade,
-  onListClick,
-  onListClose,
-}) => {
-  const { gameService } = useContext(Context);
-  const { t } = useAppTranslation();
-
-  const isListing = useSelector(gameService, _isListing);
-
-  return (
-    <>
-      <Modal show={showListItem} onHide={!isListing ? onListClose : undefined}>
-        <Panel>
-          <TradeableListItem
-            display={display}
-            tradeable={tradeable}
-            farmId={farmId}
-            id={id}
-            onListingMade={onListingMade}
-            onClose={onListClose}
-          />
-        </Panel>
-      </Modal>
-      <InnerPanel className="mb-1">
-        <div className="p-2">
-          <Label icon={tradeIcon} type="default" className="mb-2">
-            {t("marketplace.listings")}
-          </Label>
-          <div className="mb-2">
-            {!tradeable && <Loading />}
-            {tradeable?.listings.length === 0 && (
-              <p className="text-sm">{t("marketplace.noListings")}</p>
-            )}
-            {!!tradeable?.listings.length && (
-              <TradeTable
-                items={tradeable.listings.map((listing) => ({
-                  price: listing.sfl,
-                  expiresAt: "30 days", // TODO,
-                  createdById: listing.listedById,
-                }))}
-                id={farmId}
-              />
-            )}
-          </div>
-          <div className="w-full justify-end hidden sm:flex sm:visible">
-            <Button
-              className="w-full sm:w-fit"
-              disabled={!count}
-              onClick={onListClick}
-            >
-              {t("marketplace.listForSale")}
-            </Button>
-          </div>
-        </div>
-        <Button
-          className="w-full sm:hidden"
-          disabled={!count}
-          onClick={onListClick}
-        >
-          {t("marketplace.listForSale")}
-        </Button>
-      </InnerPanel>
-    </>
   );
 };
